big fixed in getsubmatrix and interface changed to the new one. Better tests.
